refactor(amf0): restructure changes
